Can click buttons now
The ears pop and you can buy modifiers to increas ppc and pot and the store works and added the timer for the ears over time and the icons are added to the store as you buy more.
I had to follow a thread, and once I started I couldn't stop until I had finished.